refactor[#193] : 재사용되는 gender , age list shared 로 경로 수정
