+++ src/features/auth/ui/UserInfoRegistrationForm.tsx
@@ -5,6 +5,7 @@ import {
   SelectOpener,
   SignUpLandingModal,
 } from "@/entities/auth/ui";
+import { ageRangeOptionList, genderOptionList } from "@/shared/constants";
 import { useModal, useSnackBar } from "@/shared/lib/overlay";
 import { useAuthStore } from "@/shared/store/auth";
 import { Badge } from "@/shared/ui/badge";
@@ -21,7 +22,6 @@ import {
   usePostDuplicateNickname,
   usePutUserInfoRegistration,
 } from "../api";
-import { ageRangeOptionList, genderOptionList } from "../constants/form";
 import { validateNickname } from "../lib";
 import { useUserInfoRegistrationFormStore } from "../store";
 import { RegionModal } from "./RegionModal";
